routes protected, landing page integrated into application
