+++ dev-desk-queue/src/components/LandingPage/LandingPage.js
@@ -1,0 +1,11 @@
+import React, { Component } from "react";
+
+export default class LandingPage extends Component {
+  render() {
+    return (
+      <div>
+        <h1>You are home</h1>
+      </div>
+    );
+  }
+}
