+++ dev-desk-queue/src/components/PrivateRoute/PrivateRoute.js
@@ -1,0 +1,23 @@
+import React from "react";
+import {Auth} from "../authentication/Authentication";
+import {Route, Redirect} from "react-router-dom";
+
+export const PrivateRoute = ({ component: Component, ...rest }) => {
+    return (
+      <Route
+        {...rest}
+        render={props =>
+          Auth.isAuthenticated ? (
+            <Component {...props} {...rest} />
+          ) : (
+            <Redirect
+              to={{
+                pathname: "/login",
+                state: { from: props.location }
+              }}
+            />
+          )
+        }
+      />
+    );
+  }
