+++ dev-desk-queue/src/index.js
@@ -1,7 +1,13 @@
 import React from "react";
 import ReactDOM from "react-dom";
+import { BrowserRouter as Router} from "react-router-dom";
 import "./index.css";
 import App from "./App";
 
+// implement withRouter HOC?
 
-ReactDOM.render(<App />, document.getElementById("root"));
+ReactDOM.render(
+	<Router>
+		<App />
+	</Router>
+, document.getElementById("root"));
